Corrected

+++ lesson1/right_triangle.rb
@@ -1,11 +1,11 @@
 puts "This program will determine type of a triangle"
 #requesting data from user
 puts "what's the A?"
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 puts "what's the B?"
-b = gets.chomp.to_i
+b = gets.chomp.to_f
 puts "what's the C?"
-c = gets.chomp.to_i
+c = gets.chomp.to_f
 
 #determining type of the triangle
 is_right = ((a**2 == b**2 + c**2) || (b**2 == a**2 + c**2) || (c**2 == a**2 + b**2))
